+++ index.d.ts
@@ -68,6 +68,19 @@ declare module 'mongoose' {
   export function connect(uri: string, callback: CallbackWithoutResult): void;
   export function connect(uri: string, options?: ConnectOptions): Promise<Mongoose>;
 
+  /**
+     * Makes the indexes in MongoDB match the indexes defined in every model's
+     * schema. This function will drop any indexes that are not defined in
+     * the model's schema except the `_id` index, and build any indexes that
+     * are in your schema but not in MongoDB.
+     */
+  export function syncIndexes(options?: Record<string, unknown>): Promise<Array<string>>;
+  export function syncIndexes(options: Record<string, unknown> | null, callback: Callback<Array<string>>): void;
+
+  /* Tells `sanitizeFilter()` to skip the given object when filtering out potential query selector injection attacks.
+   * Use this method when you have a known query selector that you want to use. */
+  export function trusted<T>(obj: T): T;
+
   /** The Mongoose module's default connection. Equivalent to `mongoose.connections[0]`, see [`connections`](#mongoose_Mongoose-connections). */
   export const connection: Connection;
 
@@ -421,6 +434,15 @@ declare module 'mongoose' {
     startSession(options: mongodb.ClientSessionOptions, cb: Callback<mongodb.ClientSession>): void;
 
     /**
+     * Makes the indexes in MongoDB match the indexes defined in every model's
+     * schema. This function will drop any indexes that are not defined in
+     * the model's schema except the `_id` index, and build any indexes that
+     * are in your schema but not in MongoDB.
+     */
+    syncIndexes(options?: Record<string, unknown>): Promise<Array<string>>;
+    syncIndexes(options: Record<string, unknown> | null, callback: Callback<Array<string>>): void;
+
+    /**
      * _Requires MongoDB >= 3.6.0._ Executes the wrapped async function
      * in a transaction. Mongoose will commit the transaction if the
      * async function executes successfully and attempt to retry if
@@ -747,9 +769,8 @@ declare module 'mongoose' {
   interface Model<T, TQueryHelpers = {}, TMethods = {}, TVirtuals = {}> extends NodeJS.EventEmitter, AcceptsDiscriminator {
     new(doc?: AnyKeys<T> & AnyObject, fields?: any | null, options?: boolean | AnyObject): HydratedDocument<T, TMethods, TVirtuals>;
 
-    aggregate<R = any>(pipeline?: any[], options?: Record<string, unknown>): Aggregate<Array<R>>;
-    aggregate<R = any>(pipeline: any[], cb: Function): Promise<Array<R>>;
-    aggregate<R = any>(pipeline: any[], options: Record<string, unknown>, cb: Function): Promise<Array<R>>;
+    aggregate<R = any>(pipeline?: PipelineStage[], options?: mongodb.AggregateOptions, callback?: Callback<R[]>): Aggregate<Array<R>>;
+    aggregate<R = any>(pipeline: PipelineStage[], cb: Function): Aggregate<Array<R>>;
 
     /** Base Mongoose instance the model uses. */
     base: typeof mongoose;
@@ -786,7 +807,7 @@ declare module 'mongoose' {
 
     /** Creates a `countDocuments` query: counts the number of documents that match `filter`. */
     countDocuments(callback?: Callback<number>): QueryWithHelpers<number, HydratedDocument<T, TMethods, TVirtuals>, TQueryHelpers, T>;
-    countDocuments(filter: FilterQuery<T>, callback?: Callback<number>): QueryWithHelpers<number, HydratedDocument<T, TMethods, TVirtuals>, TQueryHelpers, T>;
+    countDocuments(filter: FilterQuery<T>, options?: QueryOptions, callback?: Callback<number>): QueryWithHelpers<number, HydratedDocument<T, TMethods, TVirtuals>, TQueryHelpers, T>;
 
     /** Creates a new document or documents */
     create(docs: (AnyKeys<T> | AnyObject)[], options?: SaveOptions): Promise<HydratedDocument<T, TMethods, TVirtuals>[]>;
@@ -1001,7 +1022,7 @@ declare module 'mongoose' {
     replaceOne(filter?: FilterQuery<T>, replacement?: T | AnyObject, options?: QueryOptions | null, callback?: Callback): QueryWithHelpers<any, HydratedDocument<T, TMethods, TVirtuals>, TQueryHelpers, T>;
 
     /** Schema the model uses. */
-    schema: Schema;
+    schema: Schema<T>;
 
     /**
      * @deprecated use `updateOne` or `updateMany` instead.
@@ -1220,7 +1241,7 @@ declare module 'mongoose' {
     /** remove empty objects (defaults to true) */
     minimize?: boolean;
     /** if set, mongoose will call this function to allow you to transform the returned object */
-    transform?: (doc: any, ret: any, options: any) => any;
+    transform?: boolean | ((doc: any, ret: any, options: any) => any);
     /** if true, replace any conventionally populated paths with the original id in the output. Has no affect on virtual populated paths. */
     depopulate?: boolean;
     /** if false, exclude the version key (`__v` by default) from the output */
@@ -1302,7 +1323,7 @@ declare module 'mongoose' {
     methods: { [name: string]: (this: HydratedDocument<DocType, TInstanceMethods, ExtractVirtuals<M>>, ...args: any[]) => any };
 
     /** The original object passed to the schema constructor */
-    obj: any;
+    obj: SchemaDefinition<SchemaDefinitionType<DocType>>;
 
     /** Gets/sets schema paths. */
     path<ResultType extends SchemaType = SchemaType>(path: string): ResultType;
@@ -1311,7 +1332,7 @@ declare module 'mongoose' {
     /** Lists all paths and their type in the schema. */
     paths: {
       [key: string]: SchemaType;
-    }
+    };
 
     /** Returns the pathType of `path` for this schema. */
     pathType(path: string): string;
@@ -1381,14 +1402,19 @@ declare module 'mongoose' {
     virtualpath(name: string): VirtualType | null;
   }
 
+  type NumberSchemaDefinition = typeof Number | 'number' | 'Number' | typeof Schema.Types.Number;
+  type StringSchemaDefinition = typeof String | 'string' | 'String' | typeof Schema.Types.String;
+  type BooleanSchemaDefinition = typeof Boolean | 'boolean' | 'Boolean' | typeof Schema.Types.Boolean;
+  type DateSchemaDefinition = typeof NativeDate | 'date' | 'Date' | typeof Schema.Types.Date;
+
   type SchemaDefinitionWithBuiltInClass<T> = T extends number
-    ? (typeof Number | 'number' | 'Number' | typeof Schema.Types.Number)
+    ? NumberSchemaDefinition
     : T extends string
-    ? (typeof String | 'string' | 'String' | typeof Schema.Types.String)
+    ? StringSchemaDefinition
     : T extends boolean
-    ? (typeof Boolean | 'boolean' | 'Boolean' | typeof Schema.Types.Boolean)
+    ? BooleanSchemaDefinition
     : T extends NativeDate
-    ? (typeof NativeDate | 'date' | 'Date' | typeof Schema.Types.Date)
+    ? DateSchemaDefinition
     : (Function | string);
 
   type SchemaDefinitionProperty<T = undefined> = SchemaDefinitionWithBuiltInClass<T> |
@@ -1399,7 +1425,8 @@ declare module 'mongoose' {
     SchemaTypeOptions<T extends undefined ? any : T>[] |
     Function[] |
     SchemaDefinition<T> |
-    SchemaDefinition<T>[];
+    SchemaDefinition<T>[] |
+    typeof SchemaTypes.Mixed;
 
   type SchemaDefinition<T = undefined> = T extends undefined
     ? { [path: string]: SchemaDefinitionProperty; }
@@ -1436,6 +1463,10 @@ declare module 'mongoose' {
     capped?: boolean | number | { size?: number; max?: number; autoIndexId?: boolean; };
     /** Sets a default collation for every query and aggregation. */
     collation?: mongodb.CollationOptions;
+
+    /** The timeseries option to use when creating the model's collection. */
+    timeseries?: mongodb.TimeSeriesCollectionOptions;
+
     /**
      * Mongoose by default produces a collection name by passing the model name to the utils.toCollectionName
      * method. This method pluralizes the name. Set this option if you need a different name for your collection.
@@ -1564,15 +1595,18 @@ declare module 'mongoose' {
 
   export class SchemaTypeOptions<T> {
     type?:
-      T extends string | number | boolean | NativeDate | Function ? SchemaDefinitionWithBuiltInClass<T> :
-      T extends Schema<any, any, any> ? T :
+      T extends string ? StringSchemaDefinition :
+      T extends number ? NumberSchemaDefinition :
+      T extends boolean ? BooleanSchemaDefinition :
+      T extends NativeDate ? DateSchemaDefinition :
       T extends Map<any, any> ? SchemaDefinition<typeof Map> :
+      T extends Buffer ? SchemaDefinition<typeof Buffer> :
       T extends object[] ? (AnyArray<Schema<any, any, any>> | AnyArray<SchemaDefinition<Unpacked<T>>> | AnyArray<SchemaTypeOptions<Unpacked<T>>>) :
-      T extends string[] ? AnyArray<SchemaDefinitionWithBuiltInClass<string>> | AnyArray<SchemaTypeOptions<string>> :
-      T extends number[] ? AnyArray<SchemaDefinitionWithBuiltInClass<number>> | AnyArray<SchemaTypeOptions<number>> :
-      T extends boolean[] ? AnyArray<SchemaDefinitionWithBuiltInClass<boolean>> | AnyArray<SchemaTypeOptions<boolean>> :
-      T extends Function[] ? AnyArray<SchemaDefinitionWithBuiltInClass<Function>> | AnyArray<SchemaTypeOptions<Unpacked<T>>> :
-      T | typeof SchemaType | Schema<any, any, any> | SchemaDefinition<T>;
+      T extends string[] ? AnyArray<StringSchemaDefinition> | AnyArray<SchemaTypeOptions<string>> :
+      T extends number[] ? AnyArray<NumberSchemaDefinition> | AnyArray<SchemaTypeOptions<number>> :
+      T extends boolean[] ? AnyArray<BooleanSchemaDefinition> | AnyArray<SchemaTypeOptions<boolean>> :
+      T extends Function[] ? AnyArray<Function | string> | AnyArray<SchemaTypeOptions<Unpacked<T>>> :
+      T | typeof SchemaType | Schema<any, any, any> | SchemaDefinition<T> | Function | AnyArray<Function>;
 
     /** Defines a virtual with the given name that gets/sets this path. */
     alias?: string;
@@ -1655,7 +1689,7 @@ declare module 'mongoose' {
     enum?: Array<string | number | null> | ReadonlyArray<string | number | null> | { values: Array<string | number | null> | ReadonlyArray<string | number | null>, message?: string } | { [path: string]: string | number | null };
 
     /** The default [subtype](http://bsonspec.org/spec.html) associated with this buffer when it is stored in MongoDB. Only allowed for buffer paths */
-    subtype?: number
+    subtype?: number;
 
     /** The minimum value allowed for this path. Only allowed for numbers and dates. */
     min?: number | NativeDate | [number, string] | [NativeDate, string] | readonly [number, string] | readonly [NativeDate, string];
@@ -2040,7 +2074,7 @@ declare module 'mongoose' {
 
     class Decimal128 extends mongodb.Decimal128 { }
 
-    class DocumentArray<T extends Document> extends Types.Array<T> {
+    class DocumentArray<T> extends Types.Array<T> {
       /** DocumentArray constructor */
       constructor(values: any[]);
 
@@ -2152,7 +2186,7 @@ declare module 'mongoose' {
 
     /** Specifies this query as a `countDocuments` query. */
     countDocuments(callback?: Callback<number>): QueryWithHelpers<number, DocType, THelpers, RawDocType>;
-    countDocuments(criteria: FilterQuery<DocType>, callback?: Callback<number>): QueryWithHelpers<number, DocType, THelpers, RawDocType>;
+    countDocuments(criteria: FilterQuery<DocType>, options?: QueryOptions, callback?: Callback<number>): QueryWithHelpers<number, DocType, THelpers, RawDocType>;
 
     /**
      * Returns a wrapper around a [mongodb driver cursor](http://mongodb.github.io/node-mongodb-native/2.1/api/Cursor.html).
@@ -2567,16 +2601,7 @@ declare module 'mongoose' {
     [key: string]: any;
   };
 
-  type ReadonlyPartial<TSchema> = {
-    [key in keyof TSchema]?: TSchema[key];
-  };
-
-  type MatchKeysAndValues<TSchema> = ReadonlyPartial<TSchema> & AnyObject;
-
-  type ApplyBasicQueryCasting<T> = T extends mongodb.ObjectId ? T | string | (T | string)[] : // Allow strings for ObjectIds
-    T extends string ? T | RegExp | T[] : // Allow RegExps for strings
-    T extends (infer U)[] ? T | U : // Allow single array elements for arrays
-    T | T[];
+  type ApplyBasicQueryCasting<T> = T | T[] | any;
   type Condition<T> = ApplyBasicQueryCasting<T> | QuerySelector<ApplyBasicQueryCasting<T>>;
 
   type _FilterQuery<T> = {
@@ -2599,30 +2624,26 @@ declare module 'mongoose' {
     $sort?: SortValues | Record<string, SortValues>;
   };
 
-  type OnlyFieldsOfType<TSchema, FieldType = any, AssignableType = FieldType> = {
-    [key in keyof TSchema]?: [Extract<TSchema[key], FieldType>] extends [never] ? never : AssignableType;
-  };
-
   type NumericTypes = number | Decimal128 | mongodb.Double | mongodb.Int32 | mongodb.Long;
 
   type _UpdateQuery<TSchema> = {
     /** @see https://docs.mongodb.com/manual/reference/operator/update-field/ */
-    $currentDate?: OnlyFieldsOfType<TSchema, NativeDate, true | { $type: 'date' | 'timestamp' }> & AnyObject;
-    $inc?: OnlyFieldsOfType<TSchema, NumericTypes | undefined> & AnyObject;
+    $currentDate?: AnyKeys<TSchema> & AnyObject;
+    $inc?: AnyKeys<TSchema> & AnyObject;
     $min?: AnyKeys<TSchema> & AnyObject;
     $max?: AnyKeys<TSchema> & AnyObject;
-    $mul?: OnlyFieldsOfType<TSchema, NumericTypes | undefined> & AnyObject;
+    $mul?: AnyKeys<TSchema> & AnyObject;
     $rename?: { [key: string]: string };
     $set?: AnyKeys<TSchema> & AnyObject;
     $setOnInsert?: AnyKeys<TSchema> & AnyObject;
     $unset?: AnyKeys<TSchema> & AnyObject;
 
     /** @see https://docs.mongodb.com/manual/reference/operator/update-array/ */
-    $addToSet?: OnlyFieldsOfType<TSchema, any[], any> & AnyObject;
-    $pop?: OnlyFieldsOfType<TSchema, ReadonlyArray<any>, 1 | -1> & AnyObject;
-    $pull?: OnlyFieldsOfType<TSchema, ReadonlyArray<any>, any> & AnyObject;
-    $push?: OnlyFieldsOfType<TSchema, ReadonlyArray<any>, any> & AnyObject;
-    $pullAll?: OnlyFieldsOfType<TSchema, ReadonlyArray<any>, any> & AnyObject;
+    $addToSet?: AnyKeys<TSchema> & AnyObject;
+    $pop?: AnyKeys<TSchema> & AnyObject;
+    $pull?: AnyKeys<TSchema> & AnyObject;
+    $push?: AnyKeys<TSchema> & AnyObject;
+    $pullAll?: AnyKeys<TSchema> & AnyObject;
 
     /** @see https://docs.mongodb.com/manual/reference/operator/update-bitwise/ */
     $bit?: {
@@ -2642,10 +2663,10 @@ declare module 'mongoose' {
     [Extract<T, mongodb.ObjectId>] extends [never] ? T :
     T | string;
   type _UpdateQueryDef<T> = {
-    [K in keyof T]: __UpdateDefProperty<T[K]>;
+    [K in keyof T]?: __UpdateDefProperty<T[K]>;
   };
 
-  export type UpdateQuery<T> = (_UpdateQuery<_UpdateQueryDef<T>> & MatchKeysAndValues<_UpdateQueryDef<LeanDocument<T>>>);
+  export type UpdateQuery<T> = _UpdateQuery<_UpdateQueryDef<T>> & AnyObject;
 
   export type DocumentDefinition<T> = {
     [K in keyof Omit<T, Exclude<keyof Document, '_id' | 'id' | '__v'>>]:
@@ -2686,6 +2707,10 @@ declare module 'mongoose' {
 
   export type SchemaDefinitionType<T> = T extends Document ? Omit<T, Exclude<keyof Document, '_id' | 'id' | '__v'>> : T;
   export type LeanDocument<T> = Omit<_LeanDocument<T>, Exclude<keyof Document, '_id' | 'id' | '__v'> | '$isSingleNested'>;
+
+  type DeepPartial<T> = {
+    [K in keyof T]?: DeepPartial<T[K]>;
+  }
 
   export type LeanDocumentOrArray<T> = 0 extends (1 & T) ? T :
     T extends unknown[] ? LeanDocument<T[number]>[] :
@@ -2750,7 +2775,7 @@ declare module 'mongoose' {
      * Appends a new $addFields operator to this aggregate pipeline.
      * Requires MongoDB v3.4+ to work
      */
-    addFields(arg: any): this;
+    addFields(arg: PipelineStage.AddFields['$addFields']): this;
 
     /** Sets the allowDiskUse option for the aggregation query (ignored for < 2.6.0) */
     allowDiskUse(value: boolean): this;
@@ -2769,7 +2794,7 @@ declare module 'mongoose' {
     collation(options: mongodb.CollationOptions): this;
 
     /** Appends a new $count operator to this aggregate pipeline. */
-    count(countName: string): this;
+    count(countName: PipelineStage.Count['$count']): this;
 
     /**
      * Sets the cursor option for the aggregation query (ignored for < 2.6.0).
@@ -2783,13 +2808,13 @@ declare module 'mongoose' {
     explain(callback?: Callback): Promise<any>;
 
     /** Combines multiple aggregation pipelines. */
-    facet(options: any): this;
+    facet(options: PipelineStage.Facet['$facet']): this;
 
     /** Appends new custom $graphLookup operator(s) to this aggregate pipeline, performing a recursive search on a collection. */
-    graphLookup(options: any): this;
+    graphLookup(options: PipelineStage.GraphLookup['$graphLookup']): this;
 
     /** Appends new custom $group operator to this aggregate pipeline. */
-    group(arg: any): this;
+    group(arg: PipelineStage.Group['$group']): this;
 
     /** Sets the hint option for the aggregation query (ignored for < 3.6.0) */
     hint(value: Record<string, unknown> | string): this;
@@ -2798,16 +2823,16 @@ declare module 'mongoose' {
      * Appends a new $limit operator to this aggregate pipeline.
      * @param num maximum number of records to pass to the next stage
      */
-    limit(num: number): this;
+    limit(num: PipelineStage.Limit['$limit']): this;
 
     /** Appends new custom $lookup operator to this aggregate pipeline. */
-    lookup(options: any): this;
+    lookup(options: PipelineStage.Lookup['$lookup']): this;
 
     /**
      * Appends a new custom $match operator to this aggregate pipeline.
      * @param arg $match operator contents
      */
-    match(arg: any): this;
+    match(arg: PipelineStage.Match['$match']): this;
 
     /**
      * Binds this aggregate to a model.
@@ -2825,7 +2850,7 @@ declare module 'mongoose' {
     pipeline(): any[];
 
     /** Appends a new $project operator to this aggregate pipeline. */
-    project(arg: string | Object): this;
+    project(arg: PipelineStage.Project['$project']): this;
 
     /** Sets the readPreference option for the aggregation query. */
     read(pref: string | mongodb.ReadPreferenceMode, tags?: any[]): this;
@@ -2837,13 +2862,13 @@ declare module 'mongoose' {
     redact(expression: any, thenExpr: string | any, elseExpr: string | any): this;
 
     /** Appends a new $replaceRoot operator to this aggregate pipeline. */
-    replaceRoot(newRoot: object | string): this;
+    replaceRoot(newRoot: PipelineStage.ReplaceRoot['$replaceRoot']['newRoot'] | string): this;
 
     /**
      * Helper for [Atlas Text Search](https://docs.atlas.mongodb.com/reference/atlas-search/tutorial/)'s
      * `$search` stage.
      */
-    search(options: any): this;
+    search(options: PipelineStage.Search['$search']): this;
 
     /** Lets you set arbitrary options, for middleware or plugins. */
     option(value: Record<string, unknown>): this;
@@ -2861,7 +2886,7 @@ declare module 'mongoose' {
     skip(num: number): this;
 
     /** Appends a new $sort operator to this aggregate pipeline. */
-    sort(arg: any): this;
+    sort(arg: PipelineStage.Sort['$sort']): this;
 
     /** Provides promise for aggregate. */
     then: Promise<R>['then'];
@@ -2872,8 +2897,11 @@ declare module 'mongoose' {
      */
     sortByCount(arg: string | any): this;
 
+    /** Appends new $unionWith operator to this aggregate pipeline. */
+    unionWith(options: any): this;
+
     /** Appends new custom $unwind operator(s) to this aggregate pipeline. */
-    unwind(...args: any[]): this;
+    unwind(...args: PipelineStage.Unwind['$unwind'][]): this;
   }
 
   class AggregationCursor extends stream.Readable {
@@ -2911,6 +2939,273 @@ declare module 'mongoose' {
      */
     next(): Promise<any>;
     next(callback: Callback): void;
+  }
+
+  /**
+   * [Stages reference](https://docs.mongodb.com/manual/reference/operator/aggregation-pipeline/#aggregation-pipeline-stages)
+   */
+  export type PipelineStage =
+    | PipelineStage.AddFields
+    | PipelineStage.Bucket
+    | PipelineStage.BucketAuto
+    | PipelineStage.CollStats
+    | PipelineStage.Count
+    | PipelineStage.Facet
+    | PipelineStage.GeoNear
+    | PipelineStage.GraphLookup
+    | PipelineStage.Group
+    | PipelineStage.IndexStats
+    | PipelineStage.Limit
+    | PipelineStage.ListSessions
+    | PipelineStage.Lookup
+    | PipelineStage.Match
+    | PipelineStage.Merge
+    | PipelineStage.Out
+    | PipelineStage.PlanCacheStats
+    | PipelineStage.Project
+    | PipelineStage.Redact
+    | PipelineStage.ReplaceRoot
+    | PipelineStage.ReplaceWith
+    | PipelineStage.Sample
+    | PipelineStage.Search
+    | PipelineStage.Set
+    | PipelineStage.SetWindowFields
+    | PipelineStage.Skip
+    | PipelineStage.Sort
+    | PipelineStage.SortByCount
+    | PipelineStage.UnionWith
+    | PipelineStage.Unset
+    | PipelineStage.Unwind
+
+  export namespace PipelineStage {
+    export interface AddFields {
+      /** [`$addFields` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/addFields/) */
+      $addFields: Record<string, any>
+    }
+
+    export interface Bucket {
+      /** [`$bucket` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/bucket/) */
+      $bucket: {
+        groupBy: any
+        boundaries: any[]
+        default?: any
+        output?: Record<string, { [op in AccumulatorOperator]?: any }>
+      }
+    }
+
+    export interface BucketAuto {
+      /** [`$bucketAuto` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/bucketAuto/) */
+      $bucketAuto: {
+        groupBy: any
+        buckets: number
+        output?: Record<string, { [op in AccumulatorOperator]?: any }>
+        granularity?: 'R5' | 'R10' | 'R20' | 'R40' | 'R80' | '1-2-5' | 'E6' | 'E12' | 'E24' | 'E48' | 'E96' | 'E192' | 'POWERSOF2'
+      }
+    }
+
+    export interface CollStats {
+      /** [`$collStats` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/collStats/) */
+      $collStats: {
+        latencyStats?: { histograms?: boolean }
+        storageStats?: { scale?: number }
+        count?: {}
+        queryExecStats?: {}
+      }
+    }
+
+    export interface Count {
+      /** [`$count` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/count/) */
+      $count: string
+    }
+
+    export interface Facet {
+      /** [`$facet` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/facet/) */
+      $facet: Record<
+        string,
+        Exclude<PipelineStage, PipelineStage.CollStats | PipelineStage.Facet | PipelineStage.GeoNear | PipelineStage.IndexStats | PipelineStage.Out | PipelineStage.Merge | PipelineStage.PlanCacheStats>[]
+      >
+    }
+
+    export interface GeoNear {
+      /** [`$geoNear` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/geoNear/) */
+      $geoNear: {
+        near: { type: 'Point'; coordinates: [number, number] } | [number, number]
+        distanceField: string
+        distanceMultiplier?: number
+        includeLocs?: string
+        key?: string
+        maxDistance?: number
+        minDistance?: number
+        query?: AnyObject
+        spherical?: boolean
+        uniqueDocs?: boolean
+      }
+    }
+
+    export interface GraphLookup {
+      /** [`$graphLookup` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/graphLookup/) */
+      $graphLookup: {
+        from: string
+        startWith: any
+        connectFromField: string
+        connectToField: string
+        as: string
+        maxDepth?: number
+        depthField?: string
+        restrictSearchWithMatch?: AnyObject
+      }
+    }
+
+    export interface Group {
+      /** [`$group` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/group) */
+      $group: { _id: any } | { _id: any; [key: string]: { [op in AccumulatorOperator]?: any } }
+    }
+
+    export interface IndexStats {
+      /** [`$indexStats` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/indexStats/) */
+      $indexStats: {}
+    }
+
+    export interface Limit {
+      /** [`$limit` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/limit/) */
+      $limit: number
+    }
+
+    export interface ListSessions {
+      /** [`$listSessions` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/listSessions/) */
+      $listSessions: { users?: { user: string; db: string }[] } | { allUsers?: true }
+    }
+
+    export interface Lookup {
+      /** [`$lookup` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/lookup/) */
+      $lookup: {
+        from: string
+        as: string
+        localField?: string
+        foreignField?: string
+        let?: Record<string, any>
+        pipeline?: Exclude<PipelineStage, PipelineStage.Merge | PipelineStage.Out | PipelineStage.Search>[]
+      }
+    }
+
+    export interface Match {
+      /** [`$match` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/match/) */
+      $match: AnyObject
+    }
+
+    export interface Merge {
+      /** [`$merge` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/merge/) */
+      $merge: {
+        into: string | { db: string; coll: string }
+        on?: string | string[]
+        let?: Record<string, any>
+        whenMatched?: 'replace' | 'keepExisting' | 'merge' | 'fail' | Extract<PipelineStage, PipelineStage.AddFields | PipelineStage.Set | PipelineStage.Project | PipelineStage.Unset | PipelineStage.ReplaceRoot | PipelineStage.ReplaceWith>[]
+        whenNotMatched?: 'insert' | 'discard' | 'fail'
+      }
+    }
+
+    export interface Out {
+      /** [`$out` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/out/) */
+      $out: string | { db: string; coll: string }
+    }
+
+    export interface PlanCacheStats {
+      /** [`$planCacheStats` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/planCacheStats/) */
+      $planCacheStats: {}
+    }
+
+    export interface Project {
+      /** [`$project` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/project/) */
+      $project: { [field: string]: any }
+    }
+
+    export interface Redact {
+      /** [`$redact` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/redact/) */
+      $redact: any
+    }
+
+    export interface ReplaceRoot {
+      /** [`$replaceRoot` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/replaceRoot/) */
+      $replaceRoot: { newRoot: any }
+    }
+
+    export interface ReplaceWith {
+      /** [`$replaceWith` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/replaceWith/) */
+      $replaceWith: any
+    }
+
+    export interface Sample {
+      /** [`$sample` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/sample/) */
+      $sample: { size: number }
+    }
+
+    export interface Search {
+      /** [`$search` reference](https://docs.atlas.mongodb.com/reference/atlas-search/query-syntax/) */
+      $search: {
+        [key: string]: any
+        index?: string
+        highlight?: { path: string; maxCharsToExamine?: number; maxNumPassages?: number }
+      }
+    }
+
+    export interface Set {
+      /** [`$set` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/set/) */
+      $set: Record<string, any>
+    }
+
+    export interface SetWindowFields {
+      /** [`$setWindowFields` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/setWindowFields/) */
+      $setWindowFields: {
+        partitionBy?: any
+        sortBy?: Record<string, 1 | -1>
+        output: Record<
+          string,
+          { [op in WindowOperator]?: any } & {
+            window?: {
+              documents?: [string | number, string | number]
+              range?: [string | number, string | number]
+              unit?: 'year' | 'quarter' | 'month' | 'week' | 'day' | 'hour' | 'minute' | 'second' | 'millisecond'
+            }
+          }
+        >
+      }
+    }
+
+    export interface Skip {
+      /** [`$skip` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/skip/) */
+      $skip: number
+    }
+
+    export interface Sort {
+      /** [`$sort` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/sort/) */
+      $sort: Record<string, 1 | -1 | { $meta: 'textScore' }>
+    }
+
+    export interface SortByCount {
+      /** [`$sortByCount` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/sortByCount/) */
+      $sortByCount: any
+    }
+
+    export interface UnionWith {
+      /** [`$unionWith` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/unionWith/) */
+      $unionWith:
+        | string
+        | { coll: string; pipeline?: Exclude<PipelineStage, PipelineStage.Out | PipelineStage.Merge>[] }
+    }
+
+    export interface Unset {
+      /** [`$unset` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/unset/) */
+      $unset: string | string[]
+    }
+
+    export interface Unwind {
+      /** [`$unwind` reference](https://docs.mongodb.com/manual/reference/operator/aggregation/unwind/) */
+      $unwind: string | { path: string; includeArrayIndex?: string; preserveNullAndEmptyArrays?: boolean }
+    }
+
+    type AccumulatorOperator = '$accumulator' | '$addToSet' | '$avg' | '$count' | '$first' | '$last' | '$max' | '$mergeObjects' | '$min' | '$push' | '$stdDevPop' | '$stdDevSamp' | '$sum'
+
+    type WindowOperator = '$addToSet' | '$avg' | '$count' | '$covariancePop' | '$covarianceSamp' | '$derivative' | '$expMovingAvg' | '$integral' | '$max' | '$min' | '$push' | '$stdDevSamp' | '$stdDevPop' | '$sum' | '$first' | '$last' | '$shift' | '$denseRank' | '$documentNumber' | '$rank'
   }
 
   class SchemaType {